test: split operations tests
