+++ tests/operations/__init__.py
@@ -1,2 +1,26 @@
 """Tests for :mod:`barrow.operations`."""
 
+from . import (
+    test_select,
+    test_filter,
+    test_mutate,
+    test_groupby_summary,
+    test_ungroup,
+    test_env,
+    test_join,
+    test_sql,
+    test_window,
+)
+
+__all__ = [
+    "test_select",
+    "test_filter",
+    "test_mutate",
+    "test_groupby_summary",
+    "test_ungroup",
+    "test_env",
+    "test_join",
+    "test_sql",
+    "test_window",
+]
+
